+++ envolved/envvar.py
@@ -1,20 +1,16 @@
 from __future__ import annotations
 
-from dataclasses import dataclass, field
+from _weakrefset import WeakSet
 from typing import (
-    TYPE_CHECKING, Any, Callable, Dict, Generic, Iterable, List, Mapping, MutableSet, NoReturn, Optional, TypeVar,
+    Any, Callable, Dict, Iterable, MutableSet, Optional, TypeVar,
     Union, overload, Sequence
 )
 
-from _weakrefset import WeakSet
-
-from envolved.basevar import EnvVar, SchemaEnvVar, SingleEnvVar, missing
-from envolved.factory_spec import FactorySpec
-from envolved.infer_env_var import AutoTypedEnvVar, InferEnvVar, infer_type, inferred_env_var
-from envolved.utils import factory_type_hints
-
-if TYPE_CHECKING:  # pragma: no cover
-    from envolved.basevar import AsDefault, Missing
+from envolved.basevar import AsDefault, Missing
+from envolved.basevar import EnvVar, SchemaEnvVar, SingleEnvVar, missing, Discard
+from envolved.factory_spec import FactorySpec, factory_spec
+from envolved.infer_env_var import AutoTypedEnvVar, InferEnvVar, inferred_env_var
+from envolved.parsers import ParserInput
 
 T = TypeVar('T')
 
@@ -23,29 +19,30 @@ V = TypeVar('V')
 
 
 @overload
-def env_var(key: str, *, default: Union[T, Missing] = missing,
+def env_var(key: str, *, default: Union[T, Missing, AsDefault, Discard] = missing,
             description: Optional[str] = None, validators: Iterable[Callable[[T], T]] = (),
             case_sensitive: bool = False, strip_whitespaces: bool = True) -> AutoTypedEnvVar[T]:
     pass
 
 
 @overload
-def env_var(key: str, *, type: Callable[[str], T], default: Union[T, Missing] = missing,
+def env_var(key: str, *, type: ParserInput[T], default: Union[T, Missing, Discard] = missing,
             description: Optional[str] = None, validators: Iterable[Callable[[T], T]] = (),
             case_sensitive: bool = False, strip_whitespaces: bool = True) -> SingleEnvVar[T]:
     pass
 
 
 @overload
-def env_var(key: str, *, type: Callable[..., T], default: Union[T, Missing] = missing,
-            pos_args: Sequence[Union[EnvVar[Any], InferEnvVar[T]]] = (), args: Dict[str, Union[EnvVar[Any], InferEnvVar[T]]] = {},
+def env_var(key: str, *, type: Callable[..., T], default: Union[T, Missing, Discard] = missing,
+            pos_args: Sequence[Union[EnvVar[Any], InferEnvVar[T]]] = (),
+            args: Dict[str, Union[EnvVar[Any], InferEnvVar[T]]] = {},
             description: Optional[str] = None, validators: Iterable[Callable[[T], T]] = (),
-            on_partial: Union[T, Missing, AsDefault] = missing) -> SchemaEnvVar[T]:
+            on_partial: Union[T, Missing, AsDefault, Discard] = missing) -> SchemaEnvVar[T]:
     pass
 
 
-def env_var(key: str, *, type: Optional[Callable[..., T]] = None,  # type: ignore[misc]
-            default: Union[T, Missing] = missing, description: Optional[str] = None,
+def env_var(key: str, *, type: Optional[Callable[..., T]] = None,
+            default: Union[T, Missing, Discard] = missing, description: Optional[str] = None,
             validators: Iterable[Callable[[T], T]] = (), **kwargs):
     pos_args = kwargs.pop('pos_args', ())
     args = kwargs.pop('args', {})
@@ -56,20 +53,30 @@ def env_var(key: str, *, type: Optional[Callable[..., T]] = None,  # type: ignor
         on_partial = kwargs.pop('on_partial', missing)
         if kwargs:
             raise TypeError(f'Unexpected keyword arguments: {kwargs}')
+        pos = []
         keys = {}
-        factory_spec: Optional[FactorySpec] = None
+        factory_specs: Optional[FactorySpec] = None
+        for p in pos_args:
+            if isinstance(p, InferEnvVar):
+                if factory_specs is None:
+                    factory_specs = factory_spec(type)
+                idx = len(pos)
+                if idx >= len(factory_specs.positional):
+                    raise TypeError(f'Cannot infer for positional parameter {len(pos)}')
+                var_spec = factory_specs.positional[idx]
+                p = p.with_spec(idx, var_spec)
+            pos.append(p.with_prefix(key))
         for k, v in args.items():
             if isinstance(v, InferEnvVar):
-                if factory_types is None:
-                    factory_types = factory_type_hints(type)
-                var_type = factory_types.get(k)
-                if var_type is None:
+                if factory_specs is None:
+                    factory_specs = factory_spec(type)
+                var_spec = factory_specs.keyword.get(k)
+                if var_spec is None:
                     raise TypeError(f'No type hint found for parameter {k!r} in factory {type!r}')
-                v = v.with_type(var_type)
+                v = v.with_spec(k, var_spec)
             keys[k] = v.with_prefix(key)
-            assert keys[k] is not None
         ev: EnvVar = SchemaEnvVar(keys, default, type=type, on_partial=on_partial, description=description,
-                                  validators=validators)
+                                  validators=validators, pos_args=tuple(pos))
     else:
         # single var
         case_sensitive = kwargs.pop('case_sensitive', False)